Squash & refactor some commons
